css navbar et tableau fini

+++ index.php
@@ -5,22 +5,21 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" type="text/css" href="css/table.css?t=<? echo time(); ?>">
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
+    <link rel="stylesheet" type="text/css" href="css/table.css" ?t=<? echo time(); ?> <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
     <title>TOUT LES PRODUITS</title>
 </head>
 
 <body>
     <header>
-        <nav class="navbar navbar-expand-lg navbar-dark ">
+        <nav class="navbar navbar-expand-md navbar-dark ">
             <div class="container-fluid">
                 <a class="navbar-brand " href="#">LOGO</a>
                 <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNavDropdown" aria-controls="navbarNavDropdown" aria-expanded="false" aria-label="Toggle navigation">
                     <span class="navbar-toggler-icon"></span>
                 </button>
                 <div class="collapse navbar-collapse" id="navbarNavDropdown">
-                    <ul class="navbar-nav me-auto my-2 my-lg-0">
+                    <ul class="navbar-nav">
                         <li class="nav-item dropdown">
                             <a class="nav-link dropdown-toggle text-white" href="#" id="navbarDropdownMenuLink" role="button" data-bs-toggle="dropdown" aria-expanded="false">
                                 PRODUITS
@@ -48,10 +47,6 @@
                         </li>
 
                     </ul>
-                    <form class="d-flex">
-                        <input class="form-control me-2" type="search" name="s" placeholder="Référence" aria-label="Search" autocomplete="off">
-                        <button id="search" class="btn btn-outline-success" name="envoyer" type="submit">Recherche</button>
-                    </form>
                 </div>
             </div>
         </nav>
@@ -79,28 +74,31 @@
                 <th class="text-center"> Mise à jour</th>
             </tr>
         </thead>
-        <?php while ($vapoteuse = $recupvap->fetch()) { ?>
+        <?php
+        if ($recupvap->rowCount() > 0) {
+            while ($vapoteuse = $recupvap->fetch()) { ?>
 
-            <tr>
-                <th class="text-center"><?= $vapoteuse['reference']; ?></th>
-                <td><?= $vapoteuse['nom_article']; ?></td>
-                <td><?= $vapoteuse['description']; ?></td>
-                <td class="text-center"><?= $vapoteuse['prix_achat']; ?></td>
-                <td class="text-center"><?= $vapoteuse['prix_vente']; ?></td>
-                <td class="text-center"><?= $vapoteuse['quantite']; ?></td>
-                <td class="text-center"><a class="btn" href="supprimer.php?idtout=<?= $vapoteuse['id']; ?>">
-                        <button class="delete">
-                            &#128465;&#65039;
-                        </button>
-                    </a>
-                    <a class="btn" href="modifier.php?idvapmodifier=<?= $vapoteuse['id']; ?>">
-                        <button class="modifier">&#9999;&#65039;</button> </a>
-                </td>
-            </tr>
+                <tr>
+                    <th class="text-center"><?= $vapoteuse['reference']; ?></th>
+                    <td><?= $vapoteuse['nom_article']; ?></td>
+                    <td><?= $vapoteuse['description']; ?></td>
+                    <td class="text-center"><?= $vapoteuse['prix_achat']; ?></td>
+                    <td class="text-center"><?= $vapoteuse['prix_vente']; ?></td>
+                    <td class="text-center"><?= $vapoteuse['quantite']; ?></td>
+                    <td class="text-center"><a class="btn" href="supprimer.php?idtout=<?= $vapoteuse['id']; ?>">
+                            <button class="delete">
+                                &#128465;&#65039;
+                            </button>
+                        </a>
+                        <a class="btn" href="modifier.php?idvapmodifier=<?= $vapoteuse['id']; ?>">
+                            <button class="modifier">&#9999;&#65039;</button> </a>
+                    </td>
+                </tr>
 
 
 
         <?php
+            }
         }
         ?>
     </table>
